+++ MarlinKimbra4due/module/SdBaseFile.cpp
@@ -18,9 +18,10 @@
  * <http://www.gnu.org/licenses/>.
  */
 
-#include "base.h"
+#include "../base.h"
+#include "../Marlin_main.h"
+
 #if ENABLED(SDSUPPORT)
-#include "Marlin_main.h"
 
 #include "SdBaseFile.h"
 //------------------------------------------------------------------------------