+++ Documentation/changelog.md
@@ -1,3 +1,10 @@
+### Version 4.2.4
+* Added Abort on endstop hit feature
+
+### Version 4.2.3
+* Added Filament tot printed in stats
+* Overall rewrite
+
 ### Version 4.2.2
 * Added the possibility to invert the logic for lcd buttons.
 * Language files bugfix.